requirements.txt
tk
pygame
smtplib
pillow
